Rename partial names to be clearer and cleaner
