+++ portal/views/home.py
@@ -46,7 +46,6 @@ from django.views.decorators.csrf import csrf_exempt
 
 from deploy import captcha
 from portal import app_settings, email_messages
-from portal.forms.home import ContactForm
 from portal.forms.newsletter_form import NewsletterForm
 from portal.forms.play import (
     StudentLoginForm,
@@ -59,34 +58,11 @@ from portal.helpers.emails import (
     send_verification_email,
     is_verified,
     send_email,
-    CONTACT_EMAIL,
     NOTIFICATION_EMAIL,
     add_to_salesforce,
 )
 from portal.models import Teacher, Student
 from portal.permissions import logged_in_as_student, logged_in_as_teacher
-from portal.strings.about import ABOUT_BANNER
-from portal.strings.help_and_support import HELP_BANNER
-from portal.strings.play import (
-    PLAY_BANNER,
-    PLAY_BENEFITS,
-    PLAY_HEADLINE,
-    KURONO_BANNER,
-    RAPID_ROUTER_BANNER,
-)
-from portal.strings.play_aimmo import (
-    AIMMO_BENEFITS,
-    AIMMO_MAIN_HEADLINE,
-    AIMMO_PLAY_ONLINE_HEADLINE,
-    AIMMO_RESOURCES_HEADLINE,
-)
-from portal.strings.play_rapid_router import (
-    RAPID_ROUTER_BENEFITS,
-    RAPID_ROUTER_HEADLINE,
-)
-from portal.strings.privacy_policy import PRIVACY_POLICY_BANNER
-from portal.strings.teach import TEACH_BANNER, TEACH_BENEFITS
-from portal.strings.terms import TERMS_BANNER
 from portal.utils import using_two_factor
 from ratelimit.decorators import ratelimit
 
@@ -518,83 +494,6 @@ def redirect_teacher_to_correct_page(request, teacher):
         return HttpResponseRedirect(reverse_lazy("onboarding-organisation"))
 
 
-@ratelimit(
-    "ip", periods=["1m"], increment=lambda req, res: hasattr(res, "count") and res.count
-)
-def contact(request):
-    increment_count = False
-    should_use_captcha = captcha.CAPTCHA_ENABLED
-
-    anchor = ""
-
-    if request.method == "POST":
-        contact_form = ContactForm(request.POST)
-        if not should_use_captcha:
-            remove_captcha_from_forms(contact_form)
-        increment_count = True
-        if contact_form.is_valid():
-            anchor = "top"
-            email_message = email_messages.contactEmail(
-                request,
-                contact_form.cleaned_data["name"],
-                contact_form.cleaned_data["telephone"],
-                contact_form.cleaned_data["email"],
-                contact_form.cleaned_data["message"],
-                contact_form.cleaned_data["browser"],
-            )
-            send_email(
-                CONTACT_EMAIL,
-                app_settings.CONTACT_FORM_EMAILS,
-                email_message["subject"],
-                email_message["message"],
-            )
-
-            confirmed_email_message = email_messages.confirmationContactEmailMessage(
-                request,
-                contact_form.cleaned_data["name"],
-                contact_form.cleaned_data["telephone"],
-                contact_form.cleaned_data["email"],
-                contact_form.cleaned_data["message"],
-            )
-            send_email(
-                CONTACT_EMAIL,
-                [contact_form.cleaned_data["email"]],
-                confirmed_email_message["subject"],
-                confirmed_email_message["message"],
-            )
-
-            messages.success(request, "Your message was sent successfully.")
-            return render(
-                request,
-                "portal/help-and-support.html",
-                {"form": contact_form, "anchor": anchor, "BANNER": HELP_BANNER},
-            )
-        else:
-            contact_form = ContactForm(request.POST)
-            anchor = "contact"
-
-    else:
-        contact_form = ContactForm()
-
-    if not should_use_captcha:
-        remove_captcha_from_forms(contact_form)
-
-    response = render(
-        request,
-        "portal/help-and-support.html",
-        {
-            "form": contact_form,
-            "anchor": anchor,
-            "captcha": should_use_captcha,
-            "settings": app_settings,
-            "BANNER": HELP_BANNER,
-        },
-    )
-
-    response.count = increment_count
-    return response
-
-
 @csrf_exempt
 def process_newsletter_form(request):
     if request.method == "POST":
@@ -616,60 +515,3 @@ def process_newsletter_form(request):
 
 def home(request):
     return render(request, "portal/home.html")
-
-
-def play_landing_page(request):
-    return render(
-        request,
-        "portal/play.html",
-        {
-            "BANNER": PLAY_BANNER,
-            "HEADLINE": PLAY_HEADLINE,
-            "BENEFITS": PLAY_BENEFITS,
-            "RAPID_ROUTER_BANNER": RAPID_ROUTER_BANNER,
-            "KURONO_BANNER": KURONO_BANNER,
-        },
-    )
-
-
-def play_rapid_router(request):
-    return render(
-        request,
-        "portal/play_rapid-router.html",
-        {"HEADLINE": RAPID_ROUTER_HEADLINE, "BENEFITS": RAPID_ROUTER_BENEFITS},
-    )
-
-
-def play_aimmo(request):
-    return render(
-        request,
-        "portal/play_aimmo.html",
-        {
-            "BENEFITS": AIMMO_BENEFITS,
-            "HEADLINE": AIMMO_MAIN_HEADLINE,
-            "TEACHING_RESOURCES": AIMMO_RESOURCES_HEADLINE,
-            "PLAY_ONLINE": AIMMO_PLAY_ONLINE_HEADLINE,
-        },
-    )
-
-
-def teach(request):
-    return render(
-        request,
-        "portal/teach.html",
-        {"BANNER": TEACH_BANNER, "BENEFITS": TEACH_BENEFITS},
-    )
-
-
-def about(request):
-    return render(request, "portal/about.html", {"BANNER": ABOUT_BANNER})
-
-
-def terms(request):
-    return render(request, "portal/terms.html", {"BANNER": TERMS_BANNER})
-
-
-def privacy_policy(request):
-    return render(
-        request, "portal/privacy_policy.html", {"BANNER": PRIVACY_POLICY_BANNER}
-    )
