+++ portal/views/terms.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+from django.shortcuts import render
+
+from portal.strings.terms import TERMS_BANNER
+
+
+def terms(request):
+    return render(request, "portal/terms.html", {"BANNER": TERMS_BANNER})
